misc: rename rivertags packages to syntaxtags
This commit renames rivertags packages to syntaxtags, and replaces as
many references of "River" to "syntax" or "Alloy syntax" as possible.

This is a safe change for syntax/internal/rivertags as it is an internal
package and cannot be imported from the top-level module.

The tag names themselves have not yet changed, so those are left
untouched.