misc: rename binaries to Alloy
This also removes the `grafana-` prefix where present, so that
`grafana-agent` becomes `alloy`. The `grafana-` prefix was originally
added due to the ambiguity of the term "agent," but as Alloy is a much
more unique identity the prefix is no longer needed and is removed for
the sake of brevity.