misc: rename internal/flow to internal/alloy
internal/flow, referring to the Flow system, is no longer an appropriate
name and can be confusing to contributors. This commit renames it to
internal/alloy, which is a slightly better name.

However, internal/alloy is redundant with the full path
(github.com/grafana/alloy/internal/alloy), so a better name in the
future (such as controller) would be more appropriate, but would also
require a more broad refactoring of the Alloy controller.