misc: move cmd/alloy to root of repository
Following the Go documentation on how to
[Organize a Go module][modules-layout], a single-command repository
should have its main package at the root of the repository rather than
in a cmd/ folder.

As all internal tooling is now in internal/cmd, and there's only one
non-internal binary, that binary has been moved to the root of the
repository to align with best practices.

[modules-layout]: https://go.dev/doc/modules/layout