windows: use "Alloy" as service name
Previously, the Windows service used the name "Grafana Alloy," which
didn't align with the other services names (Linux, macOS) or other
release assets (Helm chart, binaries, etc).

This commit changes the service name to Alloy to be consistent across
the board with when the Grafana prefix is included to just
documentation.