packaging: properly unregister Alloy when uninstalling on Windows
grafana/agent#6550 fixed an issue where the service settings were not
being removed properly, but inadvertedly broke the process of removing
the application from the Add/Remove programs list.

Alloy is registered as a program to the 32-bit registry, and so it must
be removed from the 32-bit registry and not the 64-bit registry.

Fixes grafana/alloy#165.

As a temporary workaround to this bug, users should delete
HKLM\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\Alloy from the
32-bit registry.