main: use VERSION file as fallback for version reporting
Currently, if Alloy was built without setting build information, it
reports its version as "v0.0.0".

This commit adds using the VERSION file as fallback, so if the version
information wasn't present (such as when ussing `go run .`), the
reported version is taken from the VERSION file at the root of the
repository.

The version reported is then appended with `-devel` to indicate that it
is not the final release.

For example, as the VERSION file currently contains `v1.1.0`, then the
fallback reported version will be `v1.1.0-devel`.