ci: fix issue where creating tag files made working branch dirty (#650)
In the "Deploy to deployment_tools" pipeline, creating the `.tag-only`
file marked the working tree as dirty, which caused creating the
`.image-tag` file to report the SHA as `wip`. This then deployed the
incorrect verison.

This commit adds .tag-only and .image-tag to the .gitignore to prevent
them from being treated as marking the file tree as dirty.