internal/units: create replacement for github.com/alecthomas/units package
This commit creates a replacement for the github.com/alecthomas/units
package. A replacement is desirable because the upstream package doesn't
work very well with the Alloy syntax:

* We use the UnmarshalText implementation of Base2Bytes, which treats
  metric (MB) and IEC (MiB) suffixes as the same.
* Base2Bytes always reports values as IEC suffixes when marshalling them
  back, which can confuse users into thinking a unit conversion has
  occurred somewhere.

I think it's potentially confusing to users that setting a limit of 4MB
actually sets a limit of 4MiB, which is ~4% less than what the user
intended.

The new implementation supports parsing the same input as the old
package, including complex byte sizes such as `4MiB3KiB`, though
simplified byte sizes is preferred (`4099KiB`), and the simplified forms
are returned when marshaling back into a string.