CHANGELOG: remove date tracking from changelog
Tracking dates in the changelog provide little value and incur cost when
relases need to be delayed (such as the issues with the current 1.1
release).

This commit removes dates from changelog release sections.

As an alternative, users who wish to know the release date should use
the GitHub release publish date as the source of truth.