ci: fix bump-formula-pr action
This commit changes the bump-formula-pr action to directly call `brew
bump-formula-pr` for creating a PR to bump a formula.

The previous action didn't work for GitHub Apps as it used a user-only
API for determining the commit name and email. For now, these fields are
hardcoded for the @grafana-alloybot app.

Closes #965