fix(helm): explicitly include namespace in manifests
Currently, most of the manifests omit `.metadata.namespace`.

This works fine with `helm install` / `helm upgrade`
where all the manifests without a namespace will be installed to `--namespace`
anyway.

However, `helm template` will produce the manifests as-is (with
no namespace), regardless of `--namespace`, forcing users to add the namespace manually
or use more tooling.

This change adds the namespace to all manifests.