fix: set Probe scrape timeout correctly
We discovered that the scrapeTimeout was not set to the value configured in the Probe resource when deploying a Probe with a high scrape timeout today.

After manually verifying that the Blackbox Exporter worked, we looked into the alloy source to discover that when `m.Spec.ScrapeTimeout` is set, instead of setting `cfg.ScrapeTimeout` to the value of `m.Spec.ScrapeTimeout`, the scrape interval was overwritten.

It looks like this happened by accident since the code for both scrape interval and timeout is very similar and might have been copied with this change being overlooked.

Co-authored-by: Stefan Andres <sandres@anaconda.com>