database_observability: split out sqlparser functions
Move the sqlparser functions from query_sample to a new file. This
should allow swapping the parser for a different implementation,
if needed. We're still tied to types from the underlying library,
but I think it's fine for now.

Tests are moved around while maintaining the same coverage percentage.
Now the query_sample tests are more focussed on testing the collector
logic rather than the parsing of sql queries.