loki.secretfilter: utilize fuzz testing for complex input (#2630)
* Fuzz test the processEntry method

Using three different configurations, fuzz test the processEntry
method, seeded with varied log formats.

* Simplify fuzzing of ProcessEntry

Rather than fuzzing select configs individually, combine into a single
fuzz test that generates inputs for most of the configurations and tests
them all together.

* Add fuzz testing for component config

Fuzz test the component config and executing it against some sample
log lines. The goal here is not to fuzz test Alloy's config parsing, nor
go's regex parsing, but focusing on ensuring a valid config doesn't
cause the component to crash when it is being used.

* Add fuzz testing for Gitleaks config

Test how the component handles the variety of Gitleaks configs that can
be provided to the component. Note this does not test the toml parser
itself, which is out of scope for this testing.

* Fix bugs uncovered by fuzz testing

When a Gitleaks toml file contains either an empty regex or a regex that
matches the empty string, the component would out-of-memory when
attempting to redact. This commit excludes both of these cases.

* Add testdata to validate fixed bugs

* Check secret to redact is not empty string

If a rule regex could match an empty string, replacing it could quickly
lead to significant memory usage. Since an empty string cannot
meaningfully be a secret, add a check to validate this and skip it.

Found via go fuzzing of the Gitleaks config, sample included.

* Update changelog to reflect fixes

* Reformat changelog

* Fix fuzz-go to fuzz multiple funcs in single file

* Update FuzzConfig to use constructed arguments

Rather than generating hundreds of un-used malformed configs, build the
arguments from the component parts, fuzzing each part separately.

* Prevent divide by zero errors in alloy syntax

Also include previous prior failed fuzz test to validate it doesn't
continue to fail.

* Detect when a rule matches the redaction string

Due to the iterative implementation of the secret replacement, if the
rule regex matched the replacement string, it would iterate until
running out of memory.

* Introduce short circuiting logic for intPow

Using go's math.Pow implementation as a basis, add some short circuit
logic to avoid unnecessary work when calculating int powers.

* Update unit tests to include fuzz errors and log

* Remove unnecessary fuzz tests due to unit tests

* Update Gitleaks fuzzing to generate valid TOML