database_observability: rename query_sample collector (#3036)
This change updates the name of the collector responsible for collecting table names from a given query from query_samples to query_tables.

This is to better reflect the responsibility of the component - which is just that - to map queries to the tables involved.

The name is plural, to reflect that a query can use multiple tables. However happy to make it singular if it better fits with convention.