database_observability: add query sample collector (#3062)
This change adds a new collector, query_sample, to the database_observability component. The collector is responsible for collecting query samples. Query samples enable users to see real invocations of a given query, enabling them to explore outliers and see the real numbers instead of normalised and aggregated statistics.

Note, this replaces the previous sample collector we had, as it works in a slightly different way. Previously we were collecting the sample data available in the perf_schema.events_statements_summary_by_digest table, which is controlled/updated by performance_schema. We now use the perf_schema.events_statements_history which stores a record of executed queries that we ourselves can sample.