refactor(pyroscope.write): Forward content-type header only. (#3041)
* refactor(pyroscope.write): Forward content-type header only.

We currently forward all incoming headers of the profile /ingest request
to the endpoints defined in `pyroscope.write`.

I think this is risky. We already had bugs (like #2201) and there probably
a couple more in there already, as I noticed we send on headers like
`Content-Length`.

I think we should change this and only forward `Content-Type`, before
anyone starts relying on this, as this is the only header used:

https://github.com/simonswine/pyroscope/blob/b57274a5acbfa867660938ba7e7f5171c3114d68/pkg/ingester/pyroscope/ingest_handler.go#L132

This is still technically a breaking change, but I think it should go
forward, as the stabililty of the component is only "public-preview"
after all.

Fixes #2612

* Add changelog entry