database_observability: improve query tables parsing fallback (#3300)
Followup on a previous change to improve the parsing of query tables.
In some cases we can't detect that the sample_text is truncated with
dots, so when doing the actual parsing -if it fails- we should attempt
to parse the digest_text as a fallback.

Followup of https://github.com/grafana/alloy/pull/3281