database_observability: collect sql_text in query sample (#3254)
This change enables optional opt-in support for extracting the sql_text column as part of a query sample result.

The sql_text column provides the un-redacted query - containing all real variable values and no truncation. This is desirable to capture in certain scenarios in order to pin point exactly which values are responsible for performance characteristics.

This behaviour is disabled by default - which is the safe default behaviour to not extract (potentially) sensitive values without explicit configuration to do so.