fix: Keep ebpf component loop running
If the eBPF component encounter an error, it stops the main processing
loops, which then caused issues because any config reload for alloy
would block the whole agent.

Fixes #3332

Also for long running scrapes we were blocking the component evaluation
as we are waiting for the config to be updated. I switched this to be a
buffered channel, so the config updated are unblocked.

In case of multiple updated buffered it will skip the in between updates
and go straight to the latest version.

Fixes #1371