Pyroscope fixes (#3697)
* pyroscope.ebpf: Retry after eBPF errors and buffer config updates. (#3381)

* fix: Keep ebpf component loop running

If the eBPF component encounter an error, it stops the main processing
loops, which then caused issues because any config reload for alloy
would block the whole agent.

Fixes #3332

Also for long running scrapes we were blocking the component evaluation
as we are waiting for the config to be updated. I switched this to be a
buffered channel, so the config updated are unblocked.

In case of multiple updated buffered it will skip the in between updates
and go straight to the latest version.

Fixes #1371

* Fix data racec in test

* Limit tries to start an ebpf session

* Add info logs at session start/end

* Pass context around to shutdown properly

* fix(pyroscope.scrape): godeltaprof scraping (#3542)

* fix(pyroscope.scrape): godeltaprof scraping

* cl

* update changelog

---------

Co-authored-by: Christian Simon <simon@swine.de>
Co-authored-by: Tolya Korniltsev <korniltsev.anatoly@gmail.com>