database_observability: add locks collector (#3828)
This additional collector to the database_observability component (for MySQL) allows the user to identify which query is locking another query. This is a very common cause of performance degradation, as it is usual that concurrent threads run transactions against the same table/data, but only one can be executed at once, causing a cascade of locked threads that go all the way to the application level.

This particular collector focuses on data locks, which can be very granular, at the table level or even row and column - i.e. the command UPDATE x.Name where Name = "Joe" is gonna lock only the columns and rows being filtered.

Information is gathered from:

- data_locks: contains the current data locks being held and requested;
- data_locks_waits: implements a many-to-many relationship showing which data lock requests are being blocked by another lock being held;

Co-authored-by: Clayton Cornell <131809008+clayton-cornell@users.noreply.github.com>
Co-authored-by: Cristian Greco <cristian@regolo.cc>