database_observability: use tokenizer to extract table names as fallback
Improve parsing of table names from sql statements by using go-sqllexer
as a fallback when the primary parser fails.

This means that there are chances we log a partial list of table names
when the sql statement is truncated, but we're heading towards making
alloy more resilient and mark statements as deemed unparseable.