Add skeleton database_observability.postgres component (#3945)
This adds a new database observability component for postgres. The purpose of this component will be to collect postgres query analysis data; replicating that of the mysql component.

The required collectors for that functionality will be built over time. This change introduces the initial component so that the collectors can subsequently be added.

The most basic connection_info collector is added, so we can begin sending the database_observability_connection_info metric for pg databases.

A docs page is added for the new component but the component README is not updated to reference postgres things, while we work out the small details on configuring the DB etc.

The changelog isn't updated to include this new addition, as we don't expect adoption of this component yet - however happy to add an entry if it makes sense to do so.