database_observability: move common log labels to middleware (#4217)
This change modifies how common labels are attached to log entries throughout the database_observability.mysql and database_observability.postgres components. Previously key data was passed to each collector, which then in turn appended those to the log entries.

With this change, that data is added within a middleware, removing the need from collectors to ensure it is logged. This makes it easier for us to add other common labels to all collectors in future - such as a unique server ID.