database_observability: Add support for supplying CSP information (#4180)
This change enables database_observability users to optionally supply AWS cloud provider information for the database they are monitoring. In doing so, we can more reliably extract AWS specific information such as the region and instance name, as well as identify the AWS account info.
These variables are attached to the database_observability_connection_info metric and can be used for filtering and grouping of instances for correlation.

We start with only AWS, with the option to add more in future.