feat: Sign Windows executables and installer (#4322)
* feat: Sign Windows executables

Initial changes to support signing Windows executables before they are attached to the GitHub release.

Because Windows signing has to take place on a Windows machine, this requires the build to be split into different phases. First the binaries are built, then the Windows executables are Authenticode signed, then the signed files are attached to the GitHub release.

This may require further modularisation to allow the files that go inside the installer to also be signed, depending on how flexible the `sign` tool is to recurse into the installer to sign the contents (or not).

* feat: Refactor Windows signing

Refactor into a multi-stage pipeline so that the binaries are built, signed, then the installer is built again with those files, then everything collated at the end for publishing.

This approach currently does not sign the uninstaller as we need to run build the installer on a Windows image to be able to intercept the creation of the uninstaller to sign it before it is added to the installer.

Signed-off-by: martincostello <martin@martincostello.com>

* feat: comment out block

I thought it would be a no-op, but apparently not.

* feat: Shorten Authenticode description

The same description as used by the rpm packages looks weird in the Windows UAC dialog.

Signed-off-by: martincostello <martin@martincostello.com>

* feat: Fix artifact path

`dist.temp` not `dist.tmp`.

Signed-off-by: martincostello <martin@martincostello.com>

* feat: Fix script path

Fix incorrect path to signing script.

Signed-off-by: martincostello <martin@martincostello.com>

* feat: Bypass make to rebuild installer

Bypass make to rebuild the installer so that the signed files do not get overwritten.

* feat: Remove test changes

Remove changes from testing signing was working.

* feat: Remove PR trigger

Forgot to remove from previous commit.

* feat: Update CHANGELOG

Add note to the CHANGELOG.

* feat: Use shared action

Refactor signing to use shared action and reduce number of artifacts.

* feat: Skip jobs

Skip container jobs while testing.

* feat: Fix test build

Use hard-coded version instead of `github.ref_name`.

* feat: resolve TODOs

Restore workflow to pre-testing setup.

* Restore ZIP of Windows installer

Restore the creation of the ZIP with the Windows installer, but keeping the original .exe file.

---------

Signed-off-by: martincostello <martin@martincostello.com>