database_observability: switch schema after fetching explain plan
After `explain_plans` collector runs, the connection it last used is
left in the last used schema. If another collector grabs that same
connection later on, then it may still be using the previous schema,
leading to collector queries being reported together with the rest
of the workload queries. As a simple workaround, here we switch back
the schema to `performance_schema` before returning the connection
to the pool.