Using labels.Range() instead of range, and labels.Builder helpers instead of append
These two changes seemed to be quite intertwined, so i've grouped them in
the same commit for now

Its worth noting that in labels.Builder - the labels are guaranteed to be in
order. But for labels.ScratchBuilder this is _not_ the case, and we explicitly
need to call Sort() before retrieving with Labels().