Ensure no duplicate labels in prometheus_test, and add extra test case for this
TestComprehensive actually started to fail, because the labels were
defined both out of order and with duplicates. Because the duplicate
check relies on ordering, the test passed since the validation check
incorrectly passed. Now that we're using the new API, sorting is
automatically done and so the test failed with duplication errors.