fix(prometheus.exporter.redis): pass MaxDistinctKeyGroups option to exporter
Changes the `MaxDistinctKeyGroups` configuration option to be passed
into the Redis exporter instance. Previously it was accepted in Alloy's
configuration but not passed through, which would cause it to default
to zero regardless of how Alloy was configured.

Note that this would cause a cardinality increase in generated metrics
for instances of this component with key group metrics enabled.

Signed-off-by: stegosaurus21 <a.jason.liu21@gmail.com>