Use labels.Labels API For String Label Compatability (#4455)
* Remote sort.Sort(labels), as the Labels type is inherently sorted

From what I understand, the labels.Labels type should always be sorted.
The main exception to this is the ScratchBuilder, referenced in a later
commit

* Using labels.Range() instead of range, and labels.Builder helpers instead of append

These two changes seemed to be quite intertwined, so i've grouped them in
the same commit for now

Its worth noting that in labels.Builder - the labels are guaranteed to be in
order. But for labels.ScratchBuilder this is _not_ the case, and we explicitly
need to call Sort() before retrieving with Labels().

* Use labels.Len() instead of slice len() function

* Use labels.New() to construct new instances

* Instead of nil, using labels.EmptyLabel()

Using string labels, labels.Labels is not a slice but a struct, so we
cannot assign it to a value of nil. labels.EmptyLabel() is a helper
function to mitigate this

* Apply lint

* Ensure no duplicate labels in prometheus_test, and add extra test case for this

TestComprehensive actually started to fail, because the labels were
defined both out of order and with duplicates. Because the duplicate
check relies on ordering, the test passed since the validation check
incorrectly passed. Now that we're using the new API, sorting is
automatically done and so the test failed with duplication errors.

* Remove temporary comments to self

* https://github.com/grafana/alloy/pull/4455#discussion_r2362546444 Add back empty check