Loki Write: Internal Labels
This change is to add logic so that labels that start with `__` are
filtered out, unless they are `externalLabels`. Labels that start with
`__` are used internally for things such as metadata. If a user want to
preserve the contents of one of these internal labels, then can do so
using relabelling to remove the `__`.

This change also adds tests which check various different scenarios