Loki Write: Internal Labels (#4601)
This change is to add logic so that labels that start with `__` are
filtered out. Labels that start with
`__` are used internally for things such as metadata. If a user want to
preserve the contents of one of these internal labels, then can do so
using relabelling to remove the `__`.