ci: add ai-review workflow (#4707)
Adds a generalized tool that uses ai to analyze a PR/git diff and uses a rules file to determine how to review the code in question.

As a first use case, this PR institutes a depenency update review workflow, which is particularly helpful at outlining code changes that are needed when versions change in a go.mod file.