chore: tail fixes (#4821)
* fix: select on sending line and exit signal

Before this change if we successfully read a line we would block on
sending that line on Lines channel and the check to see if we should
exit was done at the end of the loop. With this logic caller is forced
to consume in one go routine and call close in another one.

Instead we should perform the select for exit signal while we try to
send on the channel.

* remove initial call to reopen and configure reader in TailFile

* Remove unused Error field

* set location in TailFile