fix cluster drift alert
In case the job name is equal (for e.g. k8s-monitoring installation)
multiple instances of alloy get deployed into a cluster alloy-logs,
alloy-singleton which results in an alert firing for configuration drift.
This CL addresses the issue by grouping together on the app label which
matches alloy-logs and alloy-metrics in the case of k8s-monitoring.

Fixes #4945