dbo11y: stop tracking alloy's own queries in mysql and postgres
This PR introduces strategies for excluding "own queries" in mysql and
postgres:
- postgres: here we just need to drop samples (and wait events) for the
currently connected user when scraping `pg_stat_activity`. This new
behaviour is enabled by default, as it's deemed safe/recommended.
No changes to be done on `pg_stat_statements` as that is achieved
through user permissions update (docs on the website).
- mysql: here we add a new `setup_actors` collector that checks
(and optionally updates) settings for the currently connected user.
By default it just checks that `enabled=NO` and logs out if that's not
the case. Optionally, the component can auto-update this setting itself.
This behaviour is disabled by default, though: it follows the same
pattern of `setup_consumers` where two boolean settings are required to
enable the auto-update. Additionally, auto-update requires `INSERT` and
`UPDATE` permissions (docs on the website).