feat: database_observability: stop tracking own instrumentation queries
This PR introduces a change for excluding "own queries" in postgres from
`pg_stat_activity`: we need to drop samples (and wait events) for the
currently connected user. This new behaviour is enabled by default
through the setting `exclude_current_user`, as it's deemed safe /
recommended.

No changes to be done on `pg_stat_statements` as that is achieved
through user permissions update (docs on the website).